requirements.txt
dash
pandas
geopandas
plotly-express
dash-bootstrap-components
dash-bootstrap-templates 
fiona
shapely
numpy
bs4